refactor: move page components into pages folder
